tests/requirements.txt
Appium-Python-Client==2.11.1
selenium==4.11.2
pytest==7.4.3
python-dotenv==0.19.0